[Refactor] Made a introduction component and seperate styles
