+++ src/pages/review/Review.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MainHeader from "../../components/MainHeader";
+import MainHeader from "../../components/main/MainHeader";
 
 const Review = () => {
   return (
